+++ data/beaconrestapi/src/test/java/tech/pegasys/teku/beaconrestapi/AbstractMigratedBeaconHandlerTest.java
@@ -17,6 +17,7 @@ import static org.mockito.Mockito.mock;
 
 import java.util.function.IntSupplier;
 import tech.pegasys.teku.api.ChainDataProvider;
+import tech.pegasys.teku.api.ExecutionClientDataProvider;
 import tech.pegasys.teku.api.NetworkDataProvider;
 import tech.pegasys.teku.api.SyncDataProvider;
 import tech.pegasys.teku.api.ValidatorDataProvider;
@@ -49,6 +50,8 @@ public abstract class AbstractMigratedBeaconHandlerTest {
 
   protected ChainDataProvider chainDataProvider = mock(ChainDataProvider.class);
   protected final ValidatorDataProvider validatorDataProvider = mock(ValidatorDataProvider.class);
+  protected final ExecutionClientDataProvider executionClientDataProvider =
+      mock(ExecutionClientDataProvider.class);
 
   // Getting a NullPointerException from this? Call setHandler as the first thing you do. :)
   protected StubRestApiRequest request;