+++ ethereum/executionclient/src/main/java/tech/pegasys/teku/ethereum/executionclient/web3j/Web3JClient.java
@@ -18,7 +18,6 @@ import static tech.pegasys.teku.infrastructure.exceptions.ExceptionUtil.getMessa
 import java.io.IOException;
 import java.time.Duration;
 import java.util.concurrent.TimeoutException;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicLong;
 import org.web3j.protocol.Web3j;
 import org.web3j.protocol.Web3jService;
@@ -46,7 +45,6 @@ public abstract class Web3JClient {
   // available but uses a very old value to make sure we log if the first request fails
   private final AtomicLong lastError = new AtomicLong(STARTUP_LAST_ERROR_TIME);
   private boolean initialized = false;
-  private final AtomicBoolean isAvailable = new AtomicBoolean(false);
 
   protected Web3JClient(
       final EventLogger eventLog,
@@ -113,25 +111,16 @@ public abstract class Web3JClient {
       logExecutionClientError(error, couldBeAuthError);
       executionClientEventsPublisher.onAvailabilityUpdated(false);
     }
-
-    if (isAvailable.compareAndSet(true, false)) {
-      executionClientEventsPublisher.onAvailabilityUpdated(isAvailable.get());
-    }
   }
 
   protected void handleSuccess() {
     final long lastErrorTime = lastError.getAndUpdate(x -> NO_ERROR_TIME);
     if (lastErrorTime == STARTUP_LAST_ERROR_TIME) {
       eventLog.executionClientIsOnline();
-      executionClientEventsPublisher.onAvailabilityUpdated(true);
     } else if (lastErrorTime != NO_ERROR_TIME) {
       eventLog.executionClientRecovered();
-      executionClientEventsPublisher.onAvailabilityUpdated(true);
     }
-
-    if (isAvailable.compareAndSet(false, true)) {
-      executionClientEventsPublisher.onAvailabilityUpdated(isAvailable.get());
-    }
+    executionClientEventsPublisher.onAvailabilityUpdated(true);
   }
 
   public synchronized Web3jService getWeb3jService() {